refactor: dialog 비지니스 로직 훅으로 분리 (#193)
